# Use lexical_core in CSV and JSON parser (~25% faster)